Fix PUBDEV-3211: isBitIdentical was only checking half the columns for non-compatible Frames. Add an optional argument (epsilon) for floating-point values. Also fix it to work with UUID/String columns. Make frames compatible first, then use Cmp1. Remove Cmp2.
