+++ h2o-core/src/main/java/hex/ModelBuilder.java
@@ -6,10 +6,7 @@ import water.exceptions.H2OModelBuilderIllegalArgumentException;
 import water.fvec.*;
 import water.rapids.ast.prims.advmath.AstKFold;
 import water.util.*;
-import static water.util.FrameUtils.isBitIdentical;
 
-import java.io.FileWriter;
-import java.io.IOException;
 import java.lang.reflect.Method;
 import java.util.*;
