+++ h2o-core/src/test/java/water/fvec/ExportTest.java
@@ -6,7 +6,6 @@ import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 import water.*;
 import water.parser.ParseDataset;
-import water.util.FrameUtils;
 import water.util.Log;
 
 import java.io.File;
@@ -58,7 +57,7 @@ public class ExportTest extends TestUtil {
         assertTrue(target.exists());
         imported = parseFolder(folder);
         assertEquals(fr.numRows(), imported.numRows());
-        assertTrue(FrameUtils.isBitIdentical(fr, imported));
+        assertTrue(TestUtil.isBitIdentical(fr, imported));
       } finally {
         if (rebalanced != null) rebalanced.delete();
         if (imported != null) imported.delete();
