+++ h2o-core/src/test/java/water/fvec/FrameTest.java
@@ -8,7 +8,6 @@ import water.Scope;
 import water.TestUtil;
 
 import static org.junit.Assert.assertTrue;
-import water.util.FrameUtils;
 
 /**
  * Created by tomas on 3/16/16.
@@ -35,7 +34,7 @@ public class FrameTest extends TestUtil {
     try {
       x = testData.deepSlice(new Frame(rnd.vec(1)), null);
 //      y = testData.deepSlice(new Frame(rnd.vec(2)),null);
-      assertTrue(FrameUtils.isBitIdentical(subset1, x));
+      assertTrue(TestUtil.isBitIdentical(subset1, x));
 //      assertTrue(isBitIdentical(subset2,y));
     } finally {
       Scope.exit();
