+++ h2o-core/src/test/java/water/parser/ParseFolderTestBig.java
@@ -9,7 +9,6 @@ import water.TestUtil;
 import water.fvec.Frame;
 import water.fvec.NFSFileVec;
 import water.fvec.Vec;
-import water.util.FrameUtils;
 
 public class ParseFolderTestBig extends TestUtil {
   @BeforeClass static public void setup() { stall_till_cloudsize(1); }
@@ -21,7 +20,7 @@ public class ParseFolderTestBig extends TestUtil {
     try {
       k2 = parse_test_folder("datasets/parse_folder_test");
       k1 = parse_test_file  ("datasets/UCI/UCI-large/covtype/covtype.data");
-      Assert.assertTrue("parsed values do not match!", FrameUtils.isBitIdentical(k1,k2));
+      Assert.assertTrue("parsed values do not match!", TestUtil.isBitIdentical(k1,k2));
     } finally {
       if( k1 != null ) k1.delete();
       if( k2 != null ) k2.delete();
